Put all header elements into one container
In order to properly align the header logo / header links with the
title container and header image, I moved them into one flexbox
container. This way the two elements will stay vertically aligned along
the edge of the box, regardless of the width of the display.

When they were separate elements, I had them arranged the same way. Each
had an inner flexbox with padding on the left and right sides, with
a flex basis at 80% forcing there to be extra space. I incorrectly
assumed that having the same rates would space them the same, but basis
sizes elements based on remaining space.

When the page width shrinks beyond a certain point, they become
unaligned. It's less complicated when they are in the same horizontally
flexing flexbox.